Move app location and generate src-docs
