+++ requirements.txt
@@ -1,1 +1,2 @@
+ops >= 2.2.0
 Flask == 3.0.2
