+++ src-docs/app.py.md
@@ -1,0 +1,21 @@
+<!-- markdownlint-disable -->
+
+<a href="../src/app.py#L0"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+# <kbd>module</kbd> `app.py`
+Flask application which receives GitHub webhooks. 
+
+
+---
+
+<a href="../src/app.py#L33"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+## <kbd>function</kbd> `handle_github_webhook`
+
+```python
+handle_github_webhook()
+```
+
+Receive a GitHub webhook and append the payload to a file. 
+
+
