+++ src-docs/charm.py.md
@@ -1,0 +1,77 @@
+<!-- markdownlint-disable -->
+
+<a href="../src/charm.py#L0"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+# <kbd>module</kbd> `charm.py`
+Charm the service. 
+
+Refer to the following post for a quick-start guide that will help you develop a new k8s charm using the Operator Framework: 
+
+https://discourse.charmhub.io/t/4208 
+
+**Global Variables**
+---------------
+- **VALID_LOG_LEVELS**
+
+
+---
+
+## <kbd>class</kbd> `IsCharmsTemplateCharm`
+Charm the service. 
+
+<a href="../src/charm.py#L31"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+### <kbd>function</kbd> `__init__`
+
+```python
+__init__(*args: Any)
+```
+
+Construct. 
+
+
+
+**Args:**
+ 
+ - <b>`args`</b>:  Arguments passed to the CharmBase parent constructor. 
+
+
+---
+
+#### <kbd>property</kbd> app
+
+Application that this unit is part of. 
+
+---
+
+#### <kbd>property</kbd> charm_dir
+
+Root directory of the charm as it is running. 
+
+---
+
+#### <kbd>property</kbd> config
+
+A mapping containing the charm's config and current values. 
+
+---
+
+#### <kbd>property</kbd> meta
+
+Metadata of this charm. 
+
+---
+
+#### <kbd>property</kbd> model
+
+Shortcut for more simple access the model. 
+
+---
+
+#### <kbd>property</kbd> unit
+
+Unit that this execution is responsible for. 
+
+
+
+
